Se7en segment display
The code will convert the user provided single digit integer into a seven segment display. And code will also let users add or remove number of segments(sticks) of their choice from the integer entered by the user in the beginning for seven segment conversion, after doing so , the code will print the possible digits that can be formed by adding/removing number of segments.

Note : The program will not display a UI of seven segment display. Here the seven segment display is represented as a dictionary. There is a figure in the code which is commented using docstrings , the figure represents the dictionary 'dict7'. Letters near each stick represents the segment which are defined as values to keys in dict7.