Prime number or Composite number
This code lets user know whether the integer they entered in a given range is prime or composite. In this code the range is 1000 (can be changed by programmers).
Code also provides an additional feature i.e. it will also show all the factors of the integer entered by user.