Name Searching Code
The code searches out all the names as per the initial letter given by the user. Like if any user wants all the names starting with letter 'K' then the code will print a list of all the names having 'K' as the initial.

Note : This code will only work if a letter entered by user is in capital , because all the first letters of the names are in capital in this code.